New file only for edit_distance_match
