+++ soweego/linker/baseline.py
@@ -12,12 +12,10 @@ __copyright__ = 'Copyleft 2018, Hjfocs'
 import gzip
 import json
 import logging
-from collections import defaultdict
 from os import path
 from typing import Callable, Iterable, Tuple
 
 import click
-import jellyfish
 from soweego.commons import (data_gathering, target_database, text_utils,
                              url_utils)
 from soweego.importer.models.base_entity import BaseEntity
@@ -26,11 +24,6 @@ from soweego.ingestor import wikidata_bot
 from soweego.wikidata.api_requests import get_data_for_linker
 
 LOGGER = logging.getLogger(__name__)
-EDIT_DISTANCES = {
-    'jw': jellyfish.jaro_winkler,
-    'l': jellyfish.levenshtein_distance,
-    'dl': jellyfish.damerau_levenshtein_distance
-}
 WD_IO_FILENAME = 'wikidata_%s_dataset.jsonl.gz'
 
 
@@ -143,88 +136,3 @@ def similar_tokens_match(source, target, target_pid: str, tokenize: Callable[[st
                 res_tokenized = set(res.tokens.split())
                 if len(res_tokenized) > 1 and res_tokenized.issubset(tokenized):
                     yield (qid, target_pid, res.catalog_id)
-
-
-def edit_distance_match(source, target: BaseEntity, target_pid: str, metric: str, threshold: float) -> Iterable[Tuple[str, str, str, float]]:
-    """Given a source dataset ``{identifier: {string: [languages]}}``,
-    match strings having the given edit distance ``metric``
-    above the given ``threshold`` and return a dataset
-    ``[(source_id, PID, target_id, distance_score), ...]``.
-
-    Compute the distance for each ``(source, target)`` entity pair.
-    Target candidates are acquired as follows:
-    - build a query upon the most frequent source entity strings;
-    - exact strings are joined in an OR query, e.g., ``"string1" "string2"``;
-    - run the query against a database table containing indexed of target entities.
-
-    ``distance_type`` can be one of:
-
-    - ``jw``, `Jaro-Winkler <https://en.wikipedia.org/wiki/Jaro%E2%80%93Winkler_distance>`_;
-    - ``l``, `Levenshtein<https://en.wikipedia.org/wiki/Levenshtein_distance>`_;
-    - ``dl``, `Damerau-Levenshtein<https://en.wikipedia.org/wiki/Damerau%E2%80%93Levenshtein_distance>`_.
-
-    Return ``None`` if the given edit distance is not valid.
-    """
-    distance_function = EDIT_DISTANCES.get(metric)
-    if not distance_function:
-        LOGGER.error(
-            'Invalid distance_type parameter: "%s". ' +
-            'Please pick one of "jw" (Jaro-Winkler), "l" (Levenshtein) ' +
-            'or "dl" (Damerau-Levenshtein)', metric)
-        return None
-    LOGGER.info('Using %s edit distance', distance_function.__name__)
-    for entity_row in source:
-        entity = json.loads(entity_row)
-        source_id, source_strings = entity['qid'], entity['label']
-        query, most_frequent_source_strings = _build_index_query(
-            source_strings)
-        LOGGER.debug('Query: %s', query)
-        target_candidates = data_gathering.name_fulltext_search(
-            target, query)
-        if target_candidates is None:
-            LOGGER.warning('Skipping query that went wrong: %s', query)
-            continue
-        if target_candidates == {}:
-            LOGGER.info('Skipping query with no results: %s', query)
-            continue
-        # This should be a very small loop, just 1 iteration most of the time
-        for source_string in most_frequent_source_strings:
-            source_ascii, source_normalized = text_utils.normalize(
-                source_string)
-            for result in target_candidates:
-                target_string = result.name
-                target_id = result.catalog_id
-                target_ascii, target_normalized = text_utils.normalize(
-                    target_string)
-                try:
-                    distance = distance_function(
-                        source_normalized, target_normalized)
-                # Damerau-Levenshtein does not support some Unicode code points
-                except ValueError:
-                    LOGGER.warning(
-                        'Skipping unsupported string in pair: "%s", "%s"',
-                        source_normalized, target_normalized)
-                    continue
-                LOGGER.debug('Source: %s > %s > %s - Target: %s > %s > %s - Distance: %f',
-                             source_string, source_ascii, source_normalized,
-                             target_string, target_ascii, target_normalized,
-                             distance)
-                if (metric in ('l', 'dl') and distance <= threshold) or (metric == 'jw' and distance >= threshold):
-                    yield (source_id, target_pid, target_id, distance)
-                    LOGGER.debug("It's a match! %s -> %s",
-                                 source_id, target_id)
-                else:
-                    LOGGER.debug('Skipping potential match due to the threshold: %s -> %s - Threshold: %f - Distance: %f',
-                                 source_id, target_id, threshold, distance)
-
-
-def _build_index_query(source_strings):
-    query_builder = []
-    frequencies = defaultdict(list)
-    for label, languages in source_strings.items():
-        frequencies[len(languages)].append(label)
-    most_frequent = frequencies[max(frequencies.keys())]
-    for label in most_frequent:
-        # TODO experiment with different strategies
-        query_builder.append('"%s"' % label)
-    return ' '.join(query_builder), most_frequent
